pyproject.toml
[tool.poetry]
name = "biogroupbot"
version = "0.1.0"
description = ""
authors = ["rashideveloper <bubbless7456@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
python-dotenv = "^1.0.0"
aiogram = "^2.25.1"
redis = "^4.5.5"
psycopg2-binary = "^2.9.6"
django-parler = "^2.3"
pillow = "^9.5.0"
django-ckeditor-5 = "^0.2.7"
channels = "^4.0.0"
uvicorn = "^0.22.0"
django-jazzmin = "^2.6.0"
flake8 = "^6.0.0"
isort = "^5.12.0"


[tool.poetry.group.dev.dependencies]
setuptools = "^67.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
